backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
numpy>=1.26.0
openpyxl>=3.1.0
google-genai>=1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
python-dateutil>=2.8.0
